+++ pkg/reconciler/eventtypes.go
@@ -1,0 +1,141 @@
+/*
+Copyright 2019 The Knative Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package reconciler
+
+import (
+	"context"
+	"fmt"
+
+	"go.uber.org/zap"
+	"k8s.io/apimachinery/pkg/api/equality"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	eventingv1alpha1 "knative.dev/eventing/pkg/apis/eventing/v1alpha1"
+	"knative.dev/pkg/logging"
+
+	"knative.dev/sample-source/pkg/apis/samples/v1alpha1"
+	"knative.dev/sample-source/pkg/reconciler/resources"
+)
+
+func (r *Reconciler) reconcileEventTypes(ctx context.Context, src *v1alpha1.SampleSource) error {
+	current, err := r.getEventTypes(ctx, src)
+	if err != nil {
+		logging.FromContext(ctx).Error("Unable to get existing event types", zap.Error(err))
+		return err
+	}
+
+	expected, err := r.makeEventTypes(src)
+	if err != nil {
+		return err
+	}
+
+	toCreate, toDelete := r.computeDiff(current, expected)
+
+	for _, eventType := range toDelete {
+		if err = r.EventingClientSet.EventingV1alpha1().EventTypes(src.Namespace).Delete(eventType.Name, &metav1.DeleteOptions{}); err != nil {
+			logging.FromContext(ctx).Error("Error deleting eventType", zap.Any("eventType", eventType))
+			return err
+		}
+	}
+
+	for _, eventType := range toCreate {
+		if _, err = r.EventingClientSet.EventingV1alpha1().EventTypes(src.Namespace).Create(&eventType); err != nil {
+			logging.FromContext(ctx).Error("Error creating eventType", zap.Any("eventType", eventType))
+			return err
+		}
+	}
+
+	return err
+}
+
+func (r *Reconciler) getEventTypes(ctx context.Context, src *v1alpha1.SampleSource) ([]eventingv1alpha1.EventType, error) {
+	etl, err := r.EventingClientSet.EventingV1alpha1().EventTypes(src.Namespace).List(metav1.ListOptions{
+		LabelSelector: r.getLabelSelector(src).String(),
+	})
+	if err != nil {
+		logging.FromContext(ctx).Error("Unable to list event types: %v", zap.Error(err))
+		return nil, err
+	}
+	eventTypes := make([]eventingv1alpha1.EventType, 0)
+	for _, et := range etl.Items {
+		if metav1.IsControlledBy(&et, src) {
+			eventTypes = append(eventTypes, et)
+		}
+	}
+	return eventTypes, nil
+}
+
+func (r *Reconciler) makeEventTypes(src *v1alpha1.SampleSource) ([]eventingv1alpha1.EventType, error) {
+	eventTypes := make([]eventingv1alpha1.EventType, 0)
+
+	// Only create EventTypes for Broker sinks.
+	// We add this check here in case the SampleSource was changed from Broker to non-Broker sink.
+	// If so, we need to delete the existing ones, thus we return empty expected.
+	if ref := src.Spec.Sink.GetRef(); ref == nil || ref.Kind != "Broker" {
+		return eventTypes, nil
+	}
+
+	args := &resources.EventTypeArgs{
+		Src:    src,
+		Source: r.makeEventSource(src),
+	}
+	for _, apiEventType := range samplesourceEventTypes {
+		args.Type = apiEventType
+		eventType := resources.MakeEventType(args)
+		eventTypes = append(eventTypes, eventType)
+	}
+	return eventTypes, nil
+}
+
+func (r *Reconciler) computeDiff(current []eventingv1alpha1.EventType, expected []eventingv1alpha1.EventType) ([]eventingv1alpha1.EventType, []eventingv1alpha1.EventType) {
+	toCreate := make([]eventingv1alpha1.EventType, 0)
+	toDelete := make([]eventingv1alpha1.EventType, 0)
+	currentMap := asMap(current, keyFromEventType)
+	expectedMap := asMap(expected, keyFromEventType)
+
+	// Iterate over the slices instead of the maps for predictable UT expectations.
+	for _, e := range expected {
+		if c, ok := currentMap[keyFromEventType(&e)]; !ok {
+			toCreate = append(toCreate, e)
+		} else {
+			if !equality.Semantic.DeepEqual(e.Spec, c.Spec) {
+				toDelete = append(toDelete, c)
+				toCreate = append(toCreate, e)
+			}
+		}
+	}
+	// Need to check whether the current EventTypes are not in the expected map. If so, we have to delete them.
+	// This could happen if the SampleSource CO changes its broker.
+	for _, c := range current {
+		if _, ok := expectedMap[keyFromEventType(&c)]; !ok {
+			toDelete = append(toDelete, c)
+		}
+	}
+	return toCreate, toDelete
+}
+
+func asMap(eventTypes []eventingv1alpha1.EventType, keyFunc func(*eventingv1alpha1.EventType) string) map[string]eventingv1alpha1.EventType {
+	eventTypesAsMap := make(map[string]eventingv1alpha1.EventType, 0)
+	for _, eventType := range eventTypes {
+		key := keyFunc(&eventType)
+		eventTypesAsMap[key] = eventType
+	}
+	return eventTypesAsMap
+}
+
+func keyFromEventType(eventType *eventingv1alpha1.EventType) string {
+	return fmt.Sprintf("%s_%s_%s_%s", eventType.Spec.Type, eventType.Spec.Source, eventType.Spec.Schema, eventType.Spec.Broker)
+}
